refactor(Item): item 패키지 생성 및 Items의 패키지를 item으로 설정
