+++ src/main/java/domain/item/Items.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.item;
 
 import java.util.Collections;
 import java.util.List;
