+++ src/test/java/domain/item/ItemsTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.item;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
